feat(libdenolint): Add an option to print warnings in compact format
Add both command-line and programmatic options to specify either
`compact` or `pretty` (default) formats.

BREAKING CHANGE: All public methods require the optional parameters
to be specified as an object. The alternative listing the options
as separate function arguments has been removed. (Making the `signal`
argument alternating is not possible using the current NAPI.)

How to update the source code:

```diff
  denolint(projectDir, configPath,
-   scanDirs, ignorePatterns, signal)
+   { scanDirs, ignorePatterns }, signal)

  denolintSync(projectDir, configPath,
-   scanDirs, ignorePatterns)
+   { scanDirs, ignorePatterns })

  lint(fileName, sourceCode,
-   allRules, excludeRules, includeRules, signal)
+   { allRules, excludeRules, includeRules }, signal)

  lintSync(fileName, sourceCode,
-   allRules, excludeRules, includeRules)
+   { allRules, excludeRules, includeRules })
```

The interface of the command-line tool was not affected by this change.